requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
numpy==1.26.4
packaging==24.0
pandas==2.2.2
plotly==5.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
tzdata==2024.1
urllib3==2.2.1
